refactor: added benchmark module
