+++ crates/bench/src/main.rs
@@ -1,28 +1,21 @@
 mod allocator;
+mod benchmark;
 mod cli;
 mod manifest;
 mod profiler;
 mod report;
 
-use allocator::{alloc_snapshot, reset_alloc_counters};
+use crate::benchmark::macro_bench::run_macro_suite;
+use crate::benchmark::micro_bench::run_micro_suite;
+
 use cli::{Args, BenchSuite, FailOn};
-use manifest::{
-    BenchProfile, GateConfig, GateMetric, MacroSuiteConfig, MicroSuiteConfig, load_manifest,
-};
-use profiler::{cpu_pct, current_cpu_micros, current_rss_kb, percentile_from_sorted};
+use manifest::{GateConfig, GateMetric, load_manifest};
 use report::{BenchCase, BenchReport, ReleaseBaselineEntry, ReleaseBaselineIndex};
 
 use clap::Parser;
 use std::collections::HashMap;
 use std::fs;
-use std::hint::black_box;
 use std::path::{Path, PathBuf};
-use std::time::Instant;
-
-use smallvec::SmallVec;
-use spooky_config::config::{Backend, HealthCheck, LoadBalancing, RouteMatch, Upstream};
-use spooky_edge::benchmark::{ConnectionLookupBench, RouteLookupBench};
-use spooky_lb::UpstreamPool;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 enum RegressionSeverity {
@@ -42,639 +35,6 @@ struct RegressionIssue {
     warn_limit: f64,
     severe_limit: f64,
     unit: &'static str,
-}
-
-const BENCH_SAMPLES: usize = 3;
-
-#[derive(Clone, Copy)]
-struct CaseMeasurement {
-    duration_ns: u128,
-    mean_ns: f64,
-    throughput_ops_per_sec: f64,
-    alloc_calls: u64,
-    alloc_bytes: u64,
-    rss_delta_kb: u64,
-    cpu_pct: f64,
-    p50_ns: f64,
-    p95_ns: f64,
-    p99_ns: f64,
-    max_ns: f64,
-    latency_sampled: bool,
-}
-
-fn run_case_aggregate(
-    kind: &str,
-    name: &str,
-    scale: usize,
-    iterations: u64,
-    mut op: impl FnMut() -> usize,
-) -> BenchCase {
-    let warmup_iters = (iterations / 10).clamp(100, 10_000);
-    for _ in 0..warmup_iters {
-        black_box(op());
-    }
-
-    let mut measurements = Vec::with_capacity(BENCH_SAMPLES);
-    for _ in 0..BENCH_SAMPLES {
-        reset_alloc_counters();
-        let rss_before = current_rss_kb().unwrap_or(0);
-        let cpu_before = current_cpu_micros();
-        let start = Instant::now();
-        let mut sink = 0usize;
-        for _ in 0..iterations {
-            sink ^= op();
-        }
-        black_box(sink);
-        let elapsed = start.elapsed();
-        let cpu_after = current_cpu_micros();
-        let rss_after = current_rss_kb().unwrap_or(rss_before);
-        let (alloc_calls, alloc_bytes) = alloc_snapshot();
-        let mean_ns = elapsed.as_secs_f64() * 1e9 / iterations as f64;
-        let throughput = if elapsed.as_secs_f64() > 0.0 {
-            iterations as f64 / elapsed.as_secs_f64()
-        } else {
-            0.0
-        };
-        measurements.push(CaseMeasurement {
-            duration_ns: elapsed.as_nanos(),
-            mean_ns,
-            throughput_ops_per_sec: throughput,
-            alloc_calls,
-            alloc_bytes,
-            rss_delta_kb: rss_after.saturating_sub(rss_before),
-            cpu_pct: cpu_pct(cpu_before, cpu_after, elapsed.as_nanos()),
-            p50_ns: mean_ns,
-            p95_ns: mean_ns,
-            p99_ns: mean_ns,
-            max_ns: mean_ns,
-            latency_sampled: false,
-        });
-    }
-    measurements.sort_by(|left, right| {
-        left.mean_ns
-            .partial_cmp(&right.mean_ns)
-            .unwrap_or(std::cmp::Ordering::Equal)
-    });
-    let selected = measurements[measurements.len() / 2];
-
-    BenchCase {
-        kind: kind.to_string(),
-        name: name.to_string(),
-        scale,
-        iterations,
-        duration_ns: selected.duration_ns,
-        latency_ns_per_op: selected.mean_ns,
-        throughput_ops_per_sec: selected.throughput_ops_per_sec,
-        alloc_calls: selected.alloc_calls,
-        alloc_bytes: selected.alloc_bytes,
-        rss_delta_kb: selected.rss_delta_kb,
-        cpu_pct: selected.cpu_pct,
-        latency_p50_ns: selected.p50_ns,
-        latency_p95_ns: selected.p95_ns,
-        latency_p99_ns: selected.p99_ns,
-        latency_max_ns: selected.max_ns,
-        latency_sampled: selected.latency_sampled,
-    }
-}
-
-fn run_case_with_latencies(
-    kind: &str,
-    name: &str,
-    scale: usize,
-    iterations: u64,
-    mut op: impl FnMut() -> usize,
-) -> BenchCase {
-    let warmup_iters = (iterations / 10).clamp(50, 5_000);
-    for _ in 0..warmup_iters {
-        black_box(op());
-    }
-
-    let mut measurements = Vec::with_capacity(BENCH_SAMPLES);
-    for _ in 0..BENCH_SAMPLES {
-        reset_alloc_counters();
-        let rss_before = current_rss_kb().unwrap_or(0);
-        let cpu_before = current_cpu_micros();
-        let mut latencies = Vec::with_capacity(iterations.min(200_000) as usize);
-
-        let start = Instant::now();
-        let mut sink = 0usize;
-        for _ in 0..iterations {
-            let op_start = Instant::now();
-            sink ^= op();
-            latencies.push(op_start.elapsed().as_nanos());
-        }
-        black_box(sink);
-        let elapsed = start.elapsed();
-        latencies.sort_unstable();
-
-        let cpu_after = current_cpu_micros();
-        let rss_after = current_rss_kb().unwrap_or(rss_before);
-        let (alloc_calls, alloc_bytes) = alloc_snapshot();
-        let mean_ns = elapsed.as_secs_f64() * 1e9 / iterations as f64;
-        let throughput = if elapsed.as_secs_f64() > 0.0 {
-            iterations as f64 / elapsed.as_secs_f64()
-        } else {
-            0.0
-        };
-        measurements.push(CaseMeasurement {
-            duration_ns: elapsed.as_nanos(),
-            mean_ns,
-            throughput_ops_per_sec: throughput,
-            alloc_calls,
-            alloc_bytes,
-            rss_delta_kb: rss_after.saturating_sub(rss_before),
-            cpu_pct: cpu_pct(cpu_before, cpu_after, elapsed.as_nanos()),
-            p50_ns: percentile_from_sorted(&latencies, 0.50),
-            p95_ns: percentile_from_sorted(&latencies, 0.95),
-            p99_ns: percentile_from_sorted(&latencies, 0.99),
-            max_ns: latencies.last().copied().unwrap_or(0) as f64,
-            latency_sampled: true,
-        });
-    }
-    measurements.sort_by(|left, right| {
-        left.mean_ns
-            .partial_cmp(&right.mean_ns)
-            .unwrap_or(std::cmp::Ordering::Equal)
-    });
-    let selected = measurements[measurements.len() / 2];
-
-    BenchCase {
-        kind: kind.to_string(),
-        name: name.to_string(),
-        scale,
-        iterations,
-        duration_ns: selected.duration_ns,
-        latency_ns_per_op: selected.mean_ns,
-        throughput_ops_per_sec: selected.throughput_ops_per_sec,
-        alloc_calls: selected.alloc_calls,
-        alloc_bytes: selected.alloc_bytes,
-        rss_delta_kb: selected.rss_delta_kb,
-        cpu_pct: selected.cpu_pct,
-        latency_p50_ns: selected.p50_ns,
-        latency_p95_ns: selected.p95_ns,
-        latency_p99_ns: selected.p99_ns,
-        latency_max_ns: selected.max_ns,
-        latency_sampled: selected.latency_sampled,
-    }
-}
-
-fn route_iterations(scale: usize, linear: bool) -> u64 {
-    match (scale, linear) {
-        (100, false) => 300_000,
-        (1_000, false) => 200_000,
-        (10_000, false) => 100_000,
-        (100, true) => 200_000,
-        (1_000, true) => 40_000,
-        (10_000, true) => 4_000,
-        _ => 20_000,
-    }
-}
-
-fn fast_iterations(scale: usize) -> u64 {
-    match scale {
-        100 => 300_000,
-        1_000 => 200_000,
-        10_000 => 80_000,
-        _ => 50_000,
-    }
-}
-
-fn scan_iterations(scale: usize) -> u64 {
-    match scale {
-        100 => 150_000,
-        1_000 => 30_000,
-        10_000 => 3_000,
-        _ => 10_000,
-    }
-}
-
-fn lb_ch_iterations(scale: usize) -> u64 {
-    match scale {
-        100 => 300_000,
-        1_000 => 200_000,
-        10_000 => 80_000,
-        _ => 50_000,
-    }
-}
-
-fn header_collect_iterations(scale: usize, is_large_header_set: bool) -> u64 {
-    match (scale, is_large_header_set) {
-        (100, false) => 200_000,
-        (1_000, false) => 120_000,
-        (10_000, false) => 60_000,
-        (100, true) => 120_000,
-        (1_000, true) => 80_000,
-        (10_000, true) => 40_000,
-        _ => 50_000,
-    }
-}
-
-fn synth_h3_headers(count: usize) -> Vec<(Vec<u8>, Vec<u8>)> {
-    let mut out = Vec::with_capacity(count);
-    for i in 0..count {
-        let name = match i {
-            0 => b":method".to_vec(),
-            1 => b":path".to_vec(),
-            2 => b":authority".to_vec(),
-            3 => b"user-agent".to_vec(),
-            _ => format!("x-bench-{i}").into_bytes(),
-        };
-        let value = match i {
-            0 => b"GET".to_vec(),
-            1 => b"/bench".to_vec(),
-            2 => b"example.test".to_vec(),
-            _ => format!("value-{i}").into_bytes(),
-        };
-        out.push((name, value));
-    }
-    out
-}
-
-fn benchmark_h3_header_collection(scale: usize) -> Vec<BenchCase> {
-    const INLINE_HEADERS: usize = 16;
-    let small_headers = synth_h3_headers(8);
-    let large_headers = synth_h3_headers(24);
-
-    vec![
-        run_case_aggregate(
-            "micro",
-            "h3_headers_collect_vec_small",
-            scale,
-            header_collect_iterations(scale, false),
-            || {
-                let mut headers = Vec::with_capacity(small_headers.len());
-                for (name, value) in &small_headers {
-                    headers.push((name.as_slice().to_vec(), value.as_slice().to_vec()));
-                }
-                headers.len()
-            },
-        ),
-        run_case_aggregate(
-            "micro",
-            "h3_headers_collect_smallvec_small",
-            scale,
-            header_collect_iterations(scale, false),
-            || {
-                let mut headers = SmallVec::<[(Vec<u8>, Vec<u8>); INLINE_HEADERS]>::with_capacity(
-                    small_headers.len(),
-                );
-                for (name, value) in &small_headers {
-                    headers.push((name.as_slice().to_vec(), value.as_slice().to_vec()));
-                }
-                headers.len()
-            },
-        ),
-        run_case_aggregate(
-            "micro",
-            "h3_headers_collect_vec_large",
-            scale,
-            header_collect_iterations(scale, true),
-            || {
-                let mut headers = Vec::with_capacity(large_headers.len());
-                for (name, value) in &large_headers {
-                    headers.push((name.as_slice().to_vec(), value.as_slice().to_vec()));
-                }
-                headers.len()
-            },
-        ),
-        run_case_aggregate(
-            "micro",
-            "h3_headers_collect_smallvec_large",
-            scale,
-            header_collect_iterations(scale, true),
-            || {
-                let mut headers = SmallVec::<[(Vec<u8>, Vec<u8>); INLINE_HEADERS]>::with_capacity(
-                    large_headers.len(),
-                );
-                for (name, value) in &large_headers {
-                    headers.push((name.as_slice().to_vec(), value.as_slice().to_vec()));
-                }
-                headers.len()
-            },
-        ),
-    ]
-}
-
-fn benchmark_route_lookup(scale: usize) -> Vec<BenchCase> {
-    let bench = RouteLookupBench::new(scale);
-    assert_eq!(bench.indexed_hit() > 0, bench.linear_hit() > 0);
-
-    vec![
-        run_case_aggregate(
-            "micro",
-            "route_lookup_indexed_hit",
-            scale,
-            route_iterations(scale, false),
-            || bench.indexed_hit(),
-        ),
-        run_case_aggregate(
-            "micro",
-            "route_lookup_linear_hit",
-            scale,
-            route_iterations(scale, true),
-            || bench.linear_hit(),
-        ),
-        run_case_aggregate(
-            "micro",
-            "route_lookup_indexed_miss",
-            scale,
-            route_iterations(scale, false),
-            || bench.indexed_miss(),
-        ),
-    ]
-}
-
-fn build_lb_upstream(scale: usize, lb_type: &str) -> Upstream {
-    let backends = (0..scale.max(1))
-        .map(|idx| Backend {
-            id: format!("backend-{idx:05}"),
-            address: format!("127.0.0.1:{}", 10_000 + (idx % 50_000)),
-            weight: 1,
-            health_check: Some(HealthCheck {
-                path: "/health".to_string(),
-                interval: 5_000,
-                timeout_ms: 1_000,
-                failure_threshold: 3,
-                success_threshold: 2,
-                cooldown_ms: 5_000,
-            }),
-        })
-        .collect();
-
-    Upstream {
-        load_balancing: LoadBalancing {
-            lb_type: lb_type.to_string(),
-            key: None,
-        },
-        auth: Default::default(),
-        host_policy: Default::default(),
-        forwarded_headers: Default::default(),
-        tls: None,
-        route: RouteMatch {
-            host: None,
-            path_prefix: Some("/".to_string()),
-            method: None,
-        },
-        backends,
-    }
-}
-
-fn build_lb_pool(scale: usize, lb_type: &str) -> Result<UpstreamPool, String> {
-    UpstreamPool::from_upstream(&build_lb_upstream(scale, lb_type))
-        .map_err(|err| format!("failed to build LB pool '{lb_type}' for scale {scale}: {err}"))
-}
-
-fn benchmark_lb(scale: usize) -> Result<Vec<BenchCase>, String> {
-    let mut rr_pool = build_lb_pool(scale, "round-robin")?;
-    let mut random_pool = build_lb_pool(scale, "random")?;
-    let mut ch_pool = build_lb_pool(scale, "consistent-hash")?;
-
-    let keys = [
-        "user:1", "user:2", "user:3", "user:4", "user:5", "user:6", "user:7", "user:8",
-    ];
-    let mut ch_key_idx = 0usize;
-
-    Ok(vec![
-        run_case_aggregate(
-            "micro",
-            "lb_round_robin_pick",
-            scale,
-            fast_iterations(scale),
-            || rr_pool.pick("ignored").unwrap_or(usize::MAX),
-        ),
-        run_case_aggregate(
-            "micro",
-            "lb_random_pick",
-            scale,
-            fast_iterations(scale),
-            || random_pool.pick("ignored").unwrap_or(usize::MAX),
-        ),
-        run_case_aggregate(
-            "micro",
-            "lb_consistent_hash_pick",
-            scale,
-            lb_ch_iterations(scale),
-            || {
-                let key = keys[ch_key_idx & 7];
-                ch_key_idx = ch_key_idx.wrapping_add(1);
-                ch_pool.pick(key).unwrap_or(usize::MAX)
-            },
-        ),
-    ])
-}
-
-fn benchmark_connection_lookup(scale: usize) -> Vec<BenchCase> {
-    let bench = ConnectionLookupBench::new(scale);
-    assert!(bench.peer_map_hit() > 0);
-
-    vec![
-        run_case_aggregate(
-            "micro",
-            "connection_exact_lookup",
-            scale,
-            fast_iterations(scale),
-            || bench.exact_lookup(),
-        ),
-        run_case_aggregate(
-            "micro",
-            "connection_alias_lookup",
-            scale,
-            fast_iterations(scale),
-            || bench.alias_lookup(),
-        ),
-        run_case_aggregate(
-            "micro",
-            "connection_prefix_scan_miss_lookup",
-            scale,
-            scan_iterations(scale),
-            || bench.prefix_scan_miss_lookup(),
-        ),
-        run_case_aggregate(
-            "micro",
-            "connection_peer_scan_miss",
-            scale,
-            scan_iterations(scale),
-            || bench.peer_scan_miss(),
-        ),
-        run_case_aggregate(
-            "micro",
-            "connection_peer_map_hit",
-            scale,
-            fast_iterations(scale),
-            || bench.peer_map_hit(),
-        ),
-        run_case_aggregate(
-            "micro",
-            "connection_peer_map_miss",
-            scale,
-            fast_iterations(scale),
-            || bench.peer_map_miss(),
-        ),
-    ]
-}
-
-fn benchmark_macro_traffic_mix(
-    scale: usize,
-    iterations: u64,
-    mut rr_pool: UpstreamPool,
-    mut random_pool: UpstreamPool,
-    mut ch_pool: UpstreamPool,
-) -> BenchCase {
-    let route = RouteLookupBench::new(scale);
-    let conn = ConnectionLookupBench::new(scale);
-    let mut traffic_counter = 0usize;
-
-    let mut headers_small = synth_h3_headers(8);
-    let mut headers_large = synth_h3_headers(24);
-
-    run_case_with_latencies("macro", "macro_traffic_mix", scale, iterations, move || {
-        let bucket = traffic_counter % 100;
-        traffic_counter = traffic_counter.wrapping_add(1);
-
-        if bucket < 55 {
-            let route_hit = route.indexed_hit();
-            let lb = rr_pool.pick("mix-core").unwrap_or(0);
-            let conn_hit = conn.peer_map_hit();
-            route_hit ^ lb ^ conn_hit
-        } else if bucket < 75 {
-            let route_miss = route.indexed_miss();
-            let lb = random_pool.pick("mix-random").unwrap_or(0);
-            let conn_miss = conn.peer_map_miss();
-            route_miss ^ lb ^ conn_miss
-        } else if bucket < 90 {
-            let route_hit = route.linear_hit();
-            let key = if bucket & 1 == 0 {
-                "tenant:a"
-            } else {
-                "tenant:b"
-            };
-            let lb = ch_pool.pick(key).unwrap_or(0);
-            let alias = conn.alias_lookup();
-            route_hit ^ lb ^ alias
-        } else {
-            let mut collected = 0usize;
-            if bucket & 1 == 0 {
-                headers_small.rotate_left(1);
-                for (name, value) in &headers_small {
-                    collected ^= name.len() ^ value.len();
-                }
-            } else {
-                headers_large.rotate_left(1);
-                for (name, value) in &headers_large {
-                    collected ^= name.len() ^ value.len();
-                }
-            }
-            collected ^ conn.prefix_scan_miss_lookup()
-        }
-    })
-}
-
-fn benchmark_macro_long_lived_stream(
-    scale: usize,
-    iterations: u64,
-    chunks_per_stream: usize,
-    chunk_bytes: usize,
-    mut ch_pool: UpstreamPool,
-) -> BenchCase {
-    let route = RouteLookupBench::new(scale);
-    let conn = ConnectionLookupBench::new(scale);
-
-    let payload = vec![0xAB_u8; chunk_bytes.max(1)];
-    let chunks = chunks_per_stream.max(1);
-    let mut stream_counter = 0usize;
-
-    run_case_with_latencies(
-        "macro",
-        "macro_long_lived_stream",
-        scale,
-        iterations,
-        move || {
-            let stream_id = stream_counter;
-            stream_counter = stream_counter.wrapping_add(1);
-
-            let mut checksum = route.indexed_hit() ^ conn.peer_map_hit();
-            let lb = ch_pool
-                .pick(if stream_id & 1 == 0 {
-                    "stream-even"
-                } else {
-                    "stream-odd"
-                })
-                .unwrap_or(0);
-            checksum ^= lb;
-
-            for chunk_idx in 0..chunks {
-                let offset = (chunk_idx * 31 + stream_id) % payload.len();
-                checksum ^= payload[offset] as usize;
-                if chunk_idx % 8 == 0 {
-                    checksum ^= route.indexed_hit();
-                }
-                if chunk_idx % 11 == 0 {
-                    checksum ^= conn.alias_lookup();
-                }
-            }
-
-            checksum
-        },
-    )
-}
-
-fn run_micro_suite(
-    profile: &BenchProfile,
-    config: &MicroSuiteConfig,
-) -> Result<Vec<BenchCase>, String> {
-    let mut cases = Vec::new();
-    for &scale in &profile.scales {
-        cases.extend(benchmark_route_lookup(scale));
-        cases.extend(benchmark_lb(scale)?);
-        cases.extend(benchmark_connection_lookup(scale));
-        if config.include_h3_header_collection {
-            cases.extend(benchmark_h3_header_collection(scale));
-        }
-    }
-    Ok(cases)
-}
-
-fn effective_macro_scales(profile: &BenchProfile) -> Vec<usize> {
-    if !profile.macro_scales.is_empty() {
-        return profile.macro_scales.clone();
-    }
-    if let Some(first) = profile.scales.first() {
-        return vec![*first];
-    }
-    vec![100]
-}
-
-fn run_macro_suite(
-    profile: &BenchProfile,
-    config: &MacroSuiteConfig,
-) -> Result<Vec<BenchCase>, String> {
-    let mut cases = Vec::new();
-    for scale in effective_macro_scales(profile) {
-        if config.include_traffic_mix {
-            let rr = build_lb_pool(scale, "round-robin")?;
-            let random = build_lb_pool(scale, "random")?;
-            let ch = build_lb_pool(scale, "consistent-hash")?;
-            cases.push(benchmark_macro_traffic_mix(
-                scale,
-                profile.macro_traffic_mix_iterations,
-                rr,
-                random,
-                ch,
-            ));
-        }
-
-        if config.include_long_lived_stream {
-            let ch = build_lb_pool(scale, "consistent-hash")?;
-            cases.push(benchmark_macro_long_lived_stream(
-                scale,
-                profile.macro_long_lived_stream_iterations,
-                profile.macro_long_lived_stream_chunks,
-                profile.macro_long_lived_stream_chunk_bytes,
-                ch,
-            ));
-        }
-    }
-    Ok(cases)
 }
 
 fn resolve_gate_config(mut gates: GateConfig, args: &Args) -> GateConfig {
